// src/app/core/models/role.enum.ts
export enum Role {
  CITIZEN = 'CITIZEN',
  WELFARE_OFFICER = 'WELFARE_OFFICER',
  PROGRAM_MANAGER = 'PROGRAM_MANAGER',
  ADMINISTRATOR = 'ADMINISTRATOR',
  COMPLIANCE_OFFICER = 'COMPLIANCE_OFFICER',
  GOVERNMENT_AUDITOR = 'GOVERNMENT_AUDITOR'
}

export enum UserStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  SUSPENDED = 'SUSPENDED'
}